🗺️ Atlas: Splitting mcp::server Blob
