+++ src/mcp/server/mod.rs
@@ -1727,1101 +1727,9 @@ impl AletheiaMcpServer {
         self.success_json(serde_json::Value::Object(obj))
     }
 
-    /// Cursor-mode dispatch shared by `get_outgoing_edges` /
-    /// `get_incoming_edges` (Issue #3360). `incoming` selects the direction and
-    /// the tool name the cursor is bound to.
-    fn handle_adjacency_cursor(
-        &self,
-        tool: &str,
-        incoming: bool,
-        args: &serde_json::Value,
-    ) -> CallToolResult {
-        let now = time::now();
-
-        if let Some(token) = args.get("cursor").and_then(|v| v.as_str()) {
-            let payload = match self.cursors.decode(token, tool) {
-                Ok(p) => p,
-                Err(e) => return self.error_result(e),
-            };
-            let node_id = match NodeId::new(payload.filters["node_id"].as_u64().unwrap_or(0)) {
-                Ok(id) => id,
-                Err(e) => return self.invalid_argument(&e.to_string()),
-            };
-            let label = payload.filters["label"].as_str().map(str::to_string);
-            let include_vectors = payload.filters["include_vectors"]
-                .as_bool()
-                .unwrap_or(false);
-            let snapshot = (Timestamp::from(payload.svt), Timestamp::from(payload.stt));
-            return self.emit_adjacency_cursor_page(
-                tool,
-                node_id,
-                incoming,
-                &label,
-                snapshot,
-                payload.after,
-                payload.limit,
-                include_vectors,
-                payload.cid,
-                now,
-            );
-        }
-
-        let node_id = match NodeId::new(args.get("node_id").and_then(|v| v.as_u64()).unwrap_or(0)) {
-            Ok(id) => id,
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-        let label = Self::arg_str(args, "label");
-        let include_vectors = args
-            .get("include_vectors")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let limit = Self::arg_limit(args);
-        let snapshot = (now, now);
-        self.emit_adjacency_cursor_page(
-            tool,
-            node_id,
-            incoming,
-            &label,
-            snapshot,
-            None,
-            limit,
-            include_vectors,
-            String::new(),
-            now,
-        )
-    }
-
     // ========================================================================
     // Tool Implementations
     // ========================================================================
-
-    fn handle_get_node(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetNodeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        match self.db.get_node(node_id) {
-            Ok(node) => {
-                let now = time::now();
-                let response =
-                    self.node_to_response(&node, req.include_vectors.unwrap_or(false), now);
-                self.success_json(
-                    serde_json::to_value(&response)
-                        .expect("response serialization should not fail"),
-                )
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_create_node(&self, args: serde_json::Value) -> CallToolResult {
-        let req: CreateNodeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let properties = match req.properties {
-            Some(p) => match self.json_to_property_map(&p) {
-                Ok(map) => map,
-                Err(e) => return self.invalid_argument(&format!("Invalid properties: {}", e)),
-            },
-            None => PropertyMap::default(),
-        };
-
-        let valid_from = match self.parse_opt_timestamp("valid_time", &req.valid_time) {
-            Ok(v) => v,
-            Err(result) => return result,
-        };
-        let provenance = match self.parse_opt_provenance(req.provenance) {
-            Ok(p) => p,
-            Err(result) => return result,
-        };
-
-        let mut options = crate::api::transaction::WriteRequestOptions::new();
-        if let Some(valid_from) = valid_from {
-            options = options.with_valid_from(valid_from);
-        }
-        if let Some(provenance) = provenance {
-            options = options.with_provenance(provenance);
-        }
-
-        match self
-            .db
-            .create_node_with_options(&req.label, properties, options)
-        {
-            Ok(node_id) => match self.db.get_node(node_id) {
-                Ok(node) => {
-                    let now = time::now();
-                    let response = self.node_to_response(&node, true, now);
-                    self.success_json(
-                        serde_json::to_value(&response)
-                            .expect("response serialization should not fail"),
-                    )
-                }
-                Err(e) => self.db_error(e),
-            },
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_update_node(&self, args: serde_json::Value) -> CallToolResult {
-        let req: UpdateNodeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let properties = match self.json_to_property_map(&req.properties) {
-            Ok(map) => map,
-            Err(e) => return self.invalid_argument(&format!("Invalid properties: {}", e)),
-        };
-
-        let valid_from = match self.parse_opt_timestamp("valid_time", &req.valid_time) {
-            Ok(v) => v,
-            Err(result) => return result,
-        };
-        let provenance = match self.parse_opt_provenance(req.provenance) {
-            Ok(p) => p,
-            Err(result) => return result,
-        };
-
-        let mut options = crate::api::transaction::WriteRequestOptions::new();
-        if let Some(valid_from) = valid_from {
-            options = options.with_valid_from(valid_from);
-        }
-        if let Some(provenance) = provenance {
-            options = options.with_provenance(provenance);
-        }
-
-        match self
-            .db
-            .update_node_with_options(node_id, properties, options)
-        {
-            Ok(()) => match self.db.get_node(node_id) {
-                Ok(node) => {
-                    let now = time::now();
-                    let response = self.node_to_response(&node, true, now);
-                    self.success_json(
-                        serde_json::to_value(&response)
-                            .expect("response serialization should not fail"),
-                    )
-                }
-                Err(e) => self.db_error(e),
-            },
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_delete_node(&self, args: serde_json::Value) -> CallToolResult {
-        let req: DeleteNodeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let detach = req.detach.unwrap_or(false);
-
-        let valid_from = match self.parse_opt_timestamp("valid_time", &req.valid_time) {
-            Ok(v) => v,
-            Err(result) => return result,
-        };
-
-        if detach && valid_from.is_some() {
-            return self.invalid_argument(
-                "valid_time is not supported together with detach:true; cascade delete does \
-                 not support backdating. Delete the connected edges individually with \
-                 valid_time, or omit valid_time to cascade-delete at now.",
-            );
-        }
-
-        // Perform the connected-edge check and the deletion inside a single write
-        // transaction so they observe the same storage state. Splitting the count
-        // into a separate transaction (or doing it before opening one) leaves a
-        // check-then-act gap in which a concurrent writer could add an edge after
-        // the count but before the delete, silently orphaning it. Keeping both in
-        // one closure removes that cross-transaction gap (Issue #3209).
-        enum Outcome {
-            Refused { connected_edges: usize },
-            Deleted { edges_removed: usize },
-        }
-
-        let outcome = self.db.write(|tx| -> crate::core::error::Result<Outcome> {
-            // `count_connected_edges` reads the same `current` storage the
-            // transaction's edge traversal uses, and also verifies the node
-            // exists (errors propagate via `?`).
-            let connected_edges = self.db.count_connected_edges(node_id)?;
-
-            // Refuse-by-default: never report a bare success while silently
-            // orphaning edges. The caller must opt into destruction via `detach`.
-            if connected_edges > 0 && !detach {
-                return Ok(Outcome::Refused { connected_edges });
-            }
-
-            if detach {
-                // Cascade-equivalent delete: remove the node and all connected
-                // edges, reporting exactly how many edges were removed.
-                tx.delete_node_cascade(node_id)?;
-                Ok(Outcome::Deleted {
-                    edges_removed: connected_edges,
-                })
-            } else {
-                // No connected edges: a plain delete cannot orphan anything.
-                tx.delete_node_with_valid_time(node_id, valid_from)?;
-                Ok(Outcome::Deleted { edges_removed: 0 })
-            }
-        });
-
-        let outcome = match outcome {
-            Ok(o) => o,
-            Err(e) => return self.db_error(e),
-        };
-
-        match outcome {
-            // The #3209 refusal in the #3234 structured shape:
-            // `FAILED_PRECONDITION` with `details.connected_edges`, while the
-            // legacy top-level fields are preserved additively (no loss).
-            Outcome::Refused { connected_edges } => {
-                let message = format!(
-                    "Node {} has {} connected edge(s); refusing to delete. \
-                     Pass `detach: true` to delete the node and its connected edges, \
-                     or remove the edges first.",
-                    req.node_id, connected_edges
-                );
-                let mut top_level = serde_json::Map::new();
-                top_level.insert("success".to_string(), json!(false));
-                top_level.insert("node_id".to_string(), json!(req.node_id));
-                top_level.insert("connected_edges".to_string(), json!(connected_edges));
-                top_level.insert("detach_required".to_string(), json!(true));
-                Self::error_result_with_top_level(
-                    McpError::new(McpErrorCode::FailedPrecondition, message).details(json!({
-                        "node_id": req.node_id,
-                        "connected_edges": connected_edges,
-                        "detach_required": true
-                    })),
-                    top_level,
-                )
-            }
-            Outcome::Deleted { edges_removed } => self.success_json(json!({
-                "success": true,
-                "deleted_node_id": req.node_id,
-                "detached": detach,
-                "edges_removed": edges_removed
-            })),
-        }
-    }
-
-    fn handle_retract_node(&self, args: serde_json::Value) -> CallToolResult {
-        let req: RetractNodeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (see handle_delete_node).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let detach = req.detach.unwrap_or(false);
-
-        // Matching the #3221 convention: valid_time defaults to now.
-        let valid_to = match self.parse_opt_timestamp("valid_time", &req.valid_time) {
-            Ok(v) => v.unwrap_or_else(time::now),
-            Err(result) => return result,
-        };
-
-        // Perform the connected-edge check and the retraction inside a single
-        // write transaction so they observe the same storage state — no
-        // check-then-act gap for a concurrent writer to slip an edge into
-        // (same rationale as handle_delete_node, Issue #3209).
-        enum Outcome {
-            Refused { connected_edges: usize },
-            Retracted(crate::api::transaction::RetractionResult),
-        }
-
-        let outcome = self.db.write(|tx| -> crate::core::error::Result<Outcome> {
-            use crate::api::transaction::ReadOps;
-
-            // The connected-edge contract only applies to a currently-present
-            // node; an already-retracted node short-circuits below to the
-            // idempotent result (and a nonexistent one to NOT_FOUND). All
-            // reads go through the transaction itself (buffer-aware,
-            // snapshot-isolated) rather than back through `self.db`.
-            if tx.get_node(node_id).is_ok() {
-                // Enumerate DISTINCT connected edges once — the refusal
-                // count and the detach co-retraction share the same
-                // sort/dedup list, so `connected_edges` always equals what
-                // `detach: true` would retract (a self-loop appears in both
-                // adjacency directions but is one edge).
-                let mut edge_ids = tx.get_outgoing_edges(node_id)?;
-                edge_ids.extend(tx.get_incoming_edges(node_id)?);
-                edge_ids.sort_unstable();
-                edge_ids.dedup();
-                let connected_edges = edge_ids.len();
-
-                // Refuse-by-default: never report a bare success that leaves
-                // edges pointing at a retracted node. The caller must opt
-                // into co-retraction via `detach`.
-                if connected_edges > 0 && !detach {
-                    return Ok(Outcome::Refused { connected_edges });
-                }
-
-                if detach && connected_edges > 0 {
-                    // Co-retract every connected edge at the same valid time.
-                    let mut edges_retracted = 0;
-                    for edge_id in edge_ids {
-                        let edge_result = tx.retract_edge(edge_id, valid_to)?;
-                        if !edge_result.already_retracted {
-                            edges_retracted += 1;
-                        }
-                    }
-
-                    let mut result = tx.retract_node(node_id, valid_to)?;
-                    result.edges_retracted = edges_retracted;
-                    return Ok(Outcome::Retracted(result));
-                }
-            }
-
-            Ok(Outcome::Retracted(tx.retract_node(node_id, valid_to)?))
-        });
-
-        let outcome = match outcome {
-            Ok(o) => o,
-            Err(e) => return self.db_error(e),
-        };
-
-        match outcome {
-            // The refusal in the #3234 structured shape: FAILED_PRECONDITION
-            // with `details.connected_edges`, legacy top-level fields
-            // preserved additively (byte-for-byte parallel to the
-            // handle_delete_node refusal).
-            Outcome::Refused { connected_edges } => {
-                let message = format!(
-                    "Node {} has {} connected edge(s); refusing to retract. \
-                     Pass `detach: true` to retract the node and its connected edges \
-                     at the same valid time, or retract the edges first.",
-                    req.node_id, connected_edges
-                );
-                let mut top_level = serde_json::Map::new();
-                top_level.insert("success".to_string(), json!(false));
-                top_level.insert("node_id".to_string(), json!(req.node_id));
-                top_level.insert("connected_edges".to_string(), json!(connected_edges));
-                top_level.insert("detach_required".to_string(), json!(true));
-                Self::error_result_with_top_level(
-                    McpError::new(McpErrorCode::FailedPrecondition, message).details(json!({
-                        "node_id": req.node_id,
-                        "connected_edges": connected_edges,
-                        "detach_required": true
-                    })),
-                    top_level,
-                )
-            }
-            Outcome::Retracted(result) => self.success_json(json!({
-                "success": true,
-                "node_id": req.node_id,
-                "retracted": true,
-                "already_retracted": result.already_retracted,
-                "valid_from": Self::timestamp_to_rfc3339_micros(result.valid_from),
-                "valid_to": Self::timestamp_to_rfc3339_micros(result.valid_to),
-                "edges_retracted": result.edges_retracted
-            })),
-        }
-    }
-
-    fn handle_retract_edge(&self, args: serde_json::Value) -> CallToolResult {
-        let req: RetractEdgeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let edge_id = match EdgeId::new(req.edge_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (see handle_delete_node).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        // Matching the #3221 convention: valid_time defaults to now.
-        let valid_to = match self.parse_opt_timestamp("valid_time", &req.valid_time) {
-            Ok(v) => v.unwrap_or_else(time::now),
-            Err(result) => return result,
-        };
-
-        match self.db.retract_edge(edge_id, valid_to) {
-            Ok(result) => self.success_json(json!({
-                "success": true,
-                "edge_id": req.edge_id,
-                "retracted": true,
-                "already_retracted": result.already_retracted,
-                "valid_from": Self::timestamp_to_rfc3339_micros(result.valid_from),
-                "valid_to": Self::timestamp_to_rfc3339_micros(result.valid_to)
-            })),
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_delete_node_cascade(&self, args: serde_json::Value) -> CallToolResult {
-        let req: DeleteNodeCascadeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        match self.db.write(|tx| tx.delete_node_cascade(node_id)) {
-            Ok(()) => self.success_json(json!({
-                "success": true,
-                "deleted_node_id": req.node_id,
-                "cascade": true
-            })),
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    /// Cursor-mode `list_nodes` (Issue #3360): snapshot-anchored keyset paging.
-    ///
-    /// The scan is pinned to the transaction time captured on the first page
-    /// and every page is reconstructed as of that coordinate via the same
-    /// point-in-time machinery `find_nodes_at_time` uses, so concurrent writes
-    /// after the anchor are invisible: the union of all pages equals exactly a
-    /// single unbounded `list_nodes` at the anchor moment. Requires `label`
-    /// (an unlabeled list has no enumerable, ordered candidate set).
-    fn handle_list_nodes_cursor(&self, args: &serde_json::Value) -> CallToolResult {
-        let now = time::now();
-
-        // Resume path: everything discriminating is baked into the token.
-        if let Some(token) = args.get("cursor").and_then(|v| v.as_str()) {
-            let payload = match self.cursors.decode(token, "list_nodes") {
-                Ok(p) => p,
-                Err(e) => return self.error_result(e),
-            };
-            let label = payload.filters["label"].as_str().unwrap_or("").to_string();
-            let property_key = payload.filters["property_key"].as_str().map(str::to_string);
-            let property_value = match &payload.filters["property_value"] {
-                serde_json::Value::Null => None,
-                v => Some(v.clone()),
-            };
-            let include_vectors = payload.filters["include_vectors"]
-                .as_bool()
-                .unwrap_or(false);
-            let snapshot = (Timestamp::from(payload.svt), Timestamp::from(payload.stt));
-            let candidates = match self.fetch_node_candidates(
-                &label,
-                &property_key,
-                &property_value,
-                snapshot,
-            ) {
-                Ok(c) => c,
-                Err(result) => return result,
-            };
-            return self.emit_node_cursor_page(
-                "list_nodes",
-                snapshot,
-                payload.after,
-                payload.limit,
-                include_vectors,
-                payload.filters.clone(),
-                candidates,
-                payload.cid,
-                now,
-            );
-        }
-
-        // First page: validate and pin the snapshot at "now".
-        let property_key = Self::arg_str(args, "property_key");
-        let property_value = Self::arg_value(args, "property_value");
-        if property_key.is_some() != property_value.is_some() {
-            return self.invalid_argument(
-                "Both 'property_key' and 'property_value' are required together",
-            );
-        }
-        let label = match Self::arg_str(args, "label") {
-            Some(l) => l,
-            None => {
-                return self.invalid_argument(
-                    "Cursor paging requires 'label' (an unlabeled node list has no ordered, \
-                     enumerable candidate set to page over).",
-                );
-            }
-        };
-        let limit = Self::arg_limit(args);
-        let include_vectors = args
-            .get("include_vectors")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let snapshot = (now, now);
-        let candidates =
-            match self.fetch_node_candidates(&label, &property_key, &property_value, snapshot) {
-                Ok(c) => c,
-                Err(result) => return result,
-            };
-        let filters =
-            Self::node_scan_filters(&label, &property_key, &property_value, include_vectors);
-        self.emit_node_cursor_page(
-            "list_nodes",
-            snapshot,
-            None,
-            limit,
-            include_vectors,
-            filters,
-            candidates,
-            String::new(),
-            now,
-        )
-    }
-
-    fn handle_list_nodes(&self, args: serde_json::Value) -> CallToolResult {
-        // Snapshot-anchored cursor paging (Issue #3360) is a distinct path
-        // from the legacy offset paging below, which stays unchanged for
-        // backward compatibility. The cursor parameters are additive and read
-        // straight off the raw arguments (the request structs stay unchanged).
-        if Self::cursor_requested(&args) {
-            return self.handle_list_nodes_cursor(&args);
-        }
-
-        let req: ListNodesRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        // Apply resource limits. A page must be able to carry a continuation
-        // cursor, so the limit is at least 1 (limit:0 would otherwise report
-        // has_more:true with next_offset==offset, a non-progressing page that
-        // traps a paginating caller in an infinite loop).
-        let limit = req
-            .limit
-            .unwrap_or(DEFAULT_RESULT_LIMIT)
-            .clamp(1, MAX_RESULT_LIMIT);
-        let offset = req.offset.unwrap_or(0).min(MAX_PAGINATION_OFFSET);
-
-        // Validate property filter: both key and value are required together with label
-        if req.property_key.is_some() != req.property_value.is_some() {
-            return self.invalid_argument(
-                "Both 'property_key' and 'property_value' are required together",
-            );
-        }
-        if req.property_key.is_some() && req.label.is_none() {
-            return self.invalid_argument("Property filtering requires 'label' to be specified");
-        }
-
-        // One request-scoped wallclock for every entity in the response
-        // (Issue #3391).
-        let now = time::now();
-
-        // Property-based lookup: label + property_key + property_value
-        if let (Some(label), Some(prop_key), Some(prop_val)) =
-            (&req.label, &req.property_key, &req.property_value)
-        {
-            let property_value =
-                match self.json_to_property_value(prop_val) {
-                    Some(v) => v,
-                    None => return self.invalid_argument(
-                        "Unsupported property_value type. Use strings, numbers, booleans, or null.",
-                    ),
-                };
-
-            let node_ids = self
-                .db
-                .find_nodes_by_property(label, prop_key, &property_value);
-
-            // The full matching id list is already materialized, so the total
-            // is cheap to report and `has_more` is exact.
-            let total_matching = node_ids.len();
-            let include_vectors = req.include_vectors.unwrap_or(false);
-            let mut nodes = Vec::with_capacity(limit);
-            for node_id in node_ids.into_iter().skip(offset).take(limit) {
-                if let Ok(node) = self.db.get_node(node_id) {
-                    nodes.push(self.node_to_response(&node, include_vectors, now));
-                }
-            }
-
-            // `has_more`/`next_offset` are derived from the requested window
-            // (`limit`) against `total_matching`, not from `nodes.len()`: a
-            // stale property-index entry pointing at a since-deleted node is
-            // still one of the `limit` ids this page consumed, so basing
-            // `next_offset` on the (possibly smaller) resolved count would
-            // re-skip into already-consumed ids and duplicate a row on the
-            // next page.
-            let has_more = offset.saturating_add(limit) < total_matching;
-            let mut response = json!({
-                "nodes": nodes,
-                "count": nodes.len(),
-                "offset": offset,
-                "limit": limit
-            });
-            Self::attach_completeness(&mut response, offset, limit, has_more, Some(total_matching));
-            return self.success_json(response);
-        }
-
-        // Label-only scan
-        if let Some(label) = &req.label {
-            let builder = crate::query::QueryBuilder::new().scan_label(label);
-
-            // Note: We fetch offset+limit rows then skip offset. We fetch one
-            // extra row (offset+limit+1) purely to detect whether more matching
-            // nodes exist beyond this page (`has_more`) without paying for a
-            // full-scan count. Offset is capped to prevent excessive memory use.
-            match builder.limit(limit + offset + 1).execute(&self.db) {
-                Ok(results) => {
-                    // Use iterator-based approach to avoid allocating full Vec
-                    let include_vectors = req.include_vectors.unwrap_or(false);
-                    let mut nodes = Vec::with_capacity(limit);
-                    let mut skipped = 0;
-                    let mut has_more = false;
-
-                    for row_result in results {
-                        match row_result {
-                            Ok(row) => {
-                                if skipped < offset {
-                                    skipped += 1;
-                                    continue;
-                                }
-                                if let EntityResult::Node(node) = row.entity {
-                                    if nodes.len() >= limit {
-                                        // The extra (offset+limit+1)th matching
-                                        // row proves more results remain.
-                                        has_more = true;
-                                        break;
-                                    }
-                                    nodes.push(self.node_to_response(&node, include_vectors, now));
-                                }
-                            }
-                            Err(e) => return self.db_error(e),
-                        }
-                    }
-
-                    // A label scan cannot cheaply know the matching total
-                    // (that needs a full scan), so `total_matching` is omitted;
-                    // `has_more` alone carries the completeness signal.
-                    let mut response = json!({
-                        "nodes": nodes,
-                        "count": nodes.len(),
-                        "offset": offset,
-                        "limit": limit
-                    });
-                    Self::attach_completeness(&mut response, offset, nodes.len(), has_more, None);
-                    self.success_json(response)
-                }
-                Err(e) => self.db_error(e),
-            }
-        } else {
-            // Without a label filter, we cannot efficiently list all nodes
-            // Return a helpful message
-            let mut response = json!({
-                "message": "Use 'label' filter to list nodes by type, or use 'count_nodes' for total count",
-                "total_count": self.db.node_count(),
-                "nodes": [],
-                "count": 0,
-                "offset": offset,
-                "limit": limit
-            });
-            Self::attach_completeness(&mut response, offset, 0, false, None);
-            self.success_json(response)
-        }
-    }
-
-    fn handle_count_nodes(&self, args: serde_json::Value) -> CallToolResult {
-        let req: CountNodesRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        if let Some(label) = &req.label {
-            // Use QueryBuilder to count by label efficiently without collecting all rows
-            let builder = crate::query::QueryBuilder::new().scan_label(label);
-            match builder.execute(&self.db) {
-                Ok(mut results) => {
-                    // Efficiently count without allocating a Vec
-                    match results.try_fold(0usize, |acc, row| row.map(|_| acc + 1)) {
-                        Ok(count) => self.success_json(json!({"count": count, "label": label})),
-                        Err(e) => self.error_result(McpError::from_db_error(&e).with_message(
-                            format!("Error counting nodes with label '{}': {}", label, e),
-                        )),
-                    }
-                }
-                Err(e) => self.error_result(McpError::from_db_error(&e).with_message(format!(
-                    "Error executing count query for label '{}': {}",
-                    label, e
-                ))),
-            }
-        } else {
-            self.success_json(json!({"count": self.db.node_count()}))
-        }
-    }
-
-    fn handle_get_edge(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetEdgeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let edge_id = match EdgeId::new(req.edge_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        match self.db.get_edge(edge_id) {
-            Ok(edge) => {
-                let now = time::now();
-                let response =
-                    self.edge_to_response(&edge, req.include_vectors.unwrap_or(false), now);
-                self.success_json(
-                    serde_json::to_value(&response)
-                        .expect("response serialization should not fail"),
-                )
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_create_edge(&self, args: serde_json::Value) -> CallToolResult {
-        let req: CreateEdgeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let source_id = match NodeId::new(req.source_id) {
-            Ok(id) => id,
-            Err(e) => return self.invalid_argument(&format!("Invalid source_id: {}", e)),
-        };
-
-        let target_id = match NodeId::new(req.target_id) {
-            Ok(id) => id,
-            Err(e) => return self.invalid_argument(&format!("Invalid target_id: {}", e)),
-        };
-
-        let properties = match req.properties {
-            Some(p) => match self.json_to_property_map(&p) {
-                Ok(map) => map,
-                Err(e) => return self.invalid_argument(&format!("Invalid properties: {}", e)),
-            },
-            None => PropertyMap::default(),
-        };
-
-        let valid_from = match self.parse_opt_timestamp("valid_time", &req.valid_time) {
-            Ok(v) => v,
-            Err(result) => return result,
-        };
-        let provenance = match self.parse_opt_provenance(req.provenance) {
-            Ok(p) => p,
-            Err(result) => return result,
-        };
-
-        let mut options = crate::api::transaction::WriteRequestOptions::new();
-        if let Some(valid_from) = valid_from {
-            options = options.with_valid_from(valid_from);
-        }
-        if let Some(provenance) = provenance {
-            options = options.with_provenance(provenance);
-        }
-
-        match self
-            .db
-            .create_edge_with_options(source_id, target_id, &req.label, properties, options)
-        {
-            Ok(edge_id) => match self.db.get_edge(edge_id) {
-                Ok(edge) => {
-                    let now = time::now();
-                    let response = self.edge_to_response(&edge, true, now);
-                    self.success_json(
-                        serde_json::to_value(&response)
-                            .expect("response serialization should not fail"),
-                    )
-                }
-                Err(e) => self.db_error(e),
-            },
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_update_edge(&self, args: serde_json::Value) -> CallToolResult {
-        let req: UpdateEdgeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let edge_id = match EdgeId::new(req.edge_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let properties = match self.json_to_property_map(&req.properties) {
-            Ok(map) => map,
-            Err(e) => return self.invalid_argument(&format!("Invalid properties: {}", e)),
-        };
-
-        let valid_from = match self.parse_opt_timestamp("valid_time", &req.valid_time) {
-            Ok(v) => v,
-            Err(result) => return result,
-        };
-        let provenance = match self.parse_opt_provenance(req.provenance) {
-            Ok(p) => p,
-            Err(result) => return result,
-        };
-
-        let mut options = crate::api::transaction::WriteRequestOptions::new();
-        if let Some(valid_from) = valid_from {
-            options = options.with_valid_from(valid_from);
-        }
-        if let Some(provenance) = provenance {
-            options = options.with_provenance(provenance);
-        }
-
-        match self
-            .db
-            .update_edge_with_options(edge_id, properties, options)
-        {
-            Ok(()) => match self.db.get_edge(edge_id) {
-                Ok(edge) => {
-                    let now = time::now();
-                    let response = self.edge_to_response(&edge, true, now);
-                    self.success_json(
-                        serde_json::to_value(&response)
-                            .expect("response serialization should not fail"),
-                    )
-                }
-                Err(e) => self.db_error(e),
-            },
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_delete_edge(&self, args: serde_json::Value) -> CallToolResult {
-        let req: DeleteEdgeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let edge_id = match EdgeId::new(req.edge_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let valid_from = match self.parse_opt_timestamp("valid_time", &req.valid_time) {
-            Ok(v) => v,
-            Err(result) => return result,
-        };
-
-        match self
-            .db
-            .write(|tx| tx.delete_edge_with_valid_time(edge_id, valid_from))
-        {
-            Ok(()) => self.success_json(json!({
-                "success": true,
-                "deleted_edge_id": req.edge_id
-            })),
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_list_edges(&self, args: serde_json::Value) -> CallToolResult {
-        // Cursor paging (Issue #3360) is not supported here: `list_edges` does
-        // not enumerate edges (there is no global edge scan). Rather than
-        // silently ignore the flag, direct the caller to the cursor-paged
-        // adjacency tools. (No-silent-fallback culture.)
-        if Self::cursor_requested(&args) {
-            return self.error_result(
-                McpError::new(
-                    McpErrorCode::InvalidArgument,
-                    "list_edges does not enumerate edges and is not cursorable. Use \
-                     get_outgoing_edges or get_incoming_edges from a known node -- both support \
-                     snapshot-anchored cursor paging (use_cursor / cursor).",
-                )
-                .details(json!({ "cursorable_alternatives": ["get_outgoing_edges", "get_incoming_edges"] })),
-            );
-        }
-
-        let req: ListEdgesRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        // Apply resource limits
-        let limit = req
-            .limit
-            .unwrap_or(DEFAULT_RESULT_LIMIT)
-            .min(MAX_RESULT_LIMIT);
-        let offset = req.offset.unwrap_or(0);
-
-        // Edges cannot be efficiently listed without knowing source/target nodes.
-        // Provide helpful guidance to use get_outgoing_edges or get_incoming_edges.
-        let mut response = json!({
-            "message": "Use 'get_outgoing_edges' or 'get_incoming_edges' from a known node to list edges",
-            "total_count": self.db.edge_count(),
-            "edges": [],
-            "count": 0,
-            "offset": offset,
-            "limit": limit,
-            "label_filter": req.label
-        });
-        Self::attach_completeness(&mut response, offset, 0, false, None);
-        self.success_json(response)
-    }
-
-    fn handle_count_edges(&self, args: serde_json::Value) -> CallToolResult {
-        let req: CountEdgesRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        // Note: Counting by label is not efficiently supported without iterating all edges.
-        // For now, we only support total count.
-        if req.label.is_some() {
-            self.success_json(json!({
-                "message": "Counting edges by label is not supported. Use total_count instead.",
-                "total_count": self.db.edge_count(),
-                "count": null
-            }))
-        } else {
-            self.success_json(json!({"count": self.db.edge_count()}))
-        }
-    }
-
-    fn handle_get_outgoing_edges(&self, args: serde_json::Value) -> CallToolResult {
-        // Snapshot-anchored cursor paging (Issue #3360); the full-adjacency
-        // path below is unchanged for backward compatibility.
-        if Self::cursor_requested(&args) {
-            return self.handle_adjacency_cursor("get_outgoing_edges", false, &args);
-        }
-
-        let req: GetOutgoingEdgesRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let edge_ids = if let Some(label) = &req.label {
-            self.db.get_outgoing_edges_with_label(node_id, label)
-        } else {
-            self.db.get_outgoing_edges(node_id)
-        };
-
-        let include_vectors = req.include_vectors.unwrap_or(false);
-        // One request-scoped wallclock for every entity in the response
-        // (Issue #3391).
-        let now = time::now();
-        let edges: Vec<EdgeResponse> = edge_ids
-            .into_iter()
-            .filter_map(|eid| self.db.get_edge(eid).ok())
-            .map(|e| self.edge_to_response(&e, include_vectors, now))
-            .collect();
-
-        // This handler returns the complete adjacency (no limit/offset), so the
-        // result is never truncated: `has_more` is always false and
-        // `total_matching` equals the returned count.
-        let count = edges.len();
-        let mut response = json!({
-            "edges": edges,
-            "count": count
-        });
-        Self::attach_completeness(&mut response, 0, 0, false, Some(count));
-        self.success_json(response)
-    }
-
-    fn handle_get_incoming_edges(&self, args: serde_json::Value) -> CallToolResult {
-        // Snapshot-anchored cursor paging (Issue #3360); the full-adjacency
-        // path below is unchanged for backward compatibility.
-        if Self::cursor_requested(&args) {
-            return self.handle_adjacency_cursor("get_incoming_edges", true, &args);
-        }
-
-        let req: GetIncomingEdgesRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let edge_ids = self.db.get_incoming_edges(node_id);
-
-        // Filter by label if provided
-        let include_vectors = req.include_vectors.unwrap_or(false);
-        // One request-scoped wallclock for every entity in the response
-        // (Issue #3391).
-        let now = time::now();
-        let edges: Vec<EdgeResponse> = edge_ids
-            .into_iter()
-            .filter_map(|eid| self.db.get_edge(eid).ok())
-            .filter(|e| {
-                req.label
-                    .as_ref()
-                    .map(|l| self.matches_label(e.label, l))
-                    .unwrap_or(true)
-            })
-            .map(|e| self.edge_to_response(&e, include_vectors, now))
-            .collect();
-
-        // Complete adjacency (no limit/offset): never truncated, so
-        // `has_more` is always false and `total_matching` equals the count.
-        let count = edges.len();
-        let mut response = json!({
-            "edges": edges,
-            "count": count
-        });
-        Self::attach_completeness(&mut response, 0, 0, false, Some(count));
-        self.success_json(response)
-    }
 
     /// Fetch a node, optionally as of a bi-temporal coordinate.
     fn get_node_maybe_at(
@@ -2929,86 +1837,6 @@ impl AletheiaMcpServer {
         }
     }
 
-    fn handle_traverse(&self, args: serde_json::Value) -> CallToolResult {
-        // Snapshot-anchored cursor paging (Issue #3360). Unlike the id-keyset
-        // node/adjacency scans, a DFS result order is not a simple id keyset,
-        // so traverse's cursor pins the bi-temporal snapshot (making every
-        // continuation page consistent -- AC2) and continues by an internal
-        // offset over the deterministic DFS order (v1; a depth-independent
-        // keyset traversal is a documented follow-up). The offset path below
-        // is unchanged for backward compatibility.
-        if Self::cursor_requested(&args) {
-            return self.handle_traverse_cursor(&args);
-        }
-
-        let req: TraverseRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let start_id = match NodeId::new(req.start_node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let temporal = match self
-            .resolve_bitemporal_as_of(&req.as_of_valid_time, &req.as_of_transaction_time)
-        {
-            Ok(t) => t,
-            Err(result) => return result,
-        };
-
-        // Apply resource limits to prevent DoS. A page must be able to carry a
-        // continuation cursor, so the limit is at least 1 (limit:0 would
-        // otherwise report has_more:true with next_offset==offset, a
-        // non-progressing page that traps a paginating caller in a loop).
-        let depth = req.depth.unwrap_or(1).min(MAX_TRAVERSAL_DEPTH);
-        let limit = req
-            .limit
-            .unwrap_or(DEFAULT_RESULT_LIMIT)
-            .clamp(1, MAX_RESULT_LIMIT);
-        let offset = req.offset.unwrap_or(0).min(MAX_PAGINATION_OFFSET);
-        let direction = req.direction.as_deref().unwrap_or("outgoing");
-        // One request-scoped wallclock for every entity in the response
-        // (Issue #3391).
-        let now = time::now();
-
-        let (results, has_more) = self.run_traversal(
-            start_id,
-            &req.edge_label,
-            direction,
-            depth,
-            limit,
-            offset,
-            temporal,
-            req.include_vectors.unwrap_or(false),
-            now,
-        );
-
-        let count = results.len();
-        let mut response = match temporal {
-            Some((vt, tt)) => json!({
-                "results": results,
-                "count": count,
-                "as_of_valid_time": time::to_iso8601(vt),
-                "as_of_transaction_time": time::to_iso8601(tt),
-            }),
-            None => json!({
-                "results": results,
-                "count": count
-            }),
-        };
-        // The matching total would require exhausting the traversal, so
-        // `total_matching` is omitted; `has_more`/`next_offset` carry the
-        // completeness signal.
-        Self::attach_completeness(&mut response, offset, count, has_more, None);
-        self.success_json(response)
-    }
-
     /// The DFS core of `traverse`, shared by the offset path and the
     /// snapshot-anchored cursor path (Issue #3360). Returns the page of
     /// results (after skipping `offset`, taking `limit`) and whether more
@@ -3114,961 +1942,9 @@ impl AletheiaMcpServer {
         (results, has_more)
     }
 
-    /// Cursor-mode `traverse` (Issue #3360): snapshot-pinned offset
-    /// continuation. On the first page the bi-temporal snapshot is pinned
-    /// (to the request's `as_of_*` coordinate, or to "now" if none was given,
-    /// so a current-state traversal still becomes a consistent point-in-time
-    /// scan for the duration of the cursor). Every continuation re-walks the
-    /// deterministic DFS as of that pinned snapshot and skips the already-seen
-    /// prefix, so all pages reflect one consistent moment.
-    fn handle_traverse_cursor(&self, args: &serde_json::Value) -> CallToolResult {
-        let now = time::now();
-
-        // Resolve page parameters, start node, and pinned snapshot, from the
-        // token when resuming or from the request on the first page.
-        let (
-            start_id,
-            edge_label,
-            direction,
-            depth,
-            limit,
-            offset,
-            include_vectors,
-            snapshot,
-            parent_cid,
-        ) = if let Some(token) = args.get("cursor").and_then(|v| v.as_str()) {
-            let payload = match self.cursors.decode(token, "traverse") {
-                Ok(p) => p,
-                Err(e) => return self.error_result(e),
-            };
-            let f = &payload.filters;
-            let start_id = match NodeId::new(f["start_node_id"].as_u64().unwrap_or(0)) {
-                Ok(id) => id,
-                Err(e) => return self.invalid_argument(&e.to_string()),
-            };
-            (
-                start_id,
-                f["edge_label"].as_str().unwrap_or("").to_string(),
-                f["direction"].as_str().unwrap_or("outgoing").to_string(),
-                f["depth"].as_u64().unwrap_or(1) as usize,
-                payload.limit,
-                payload.off as usize,
-                f["include_vectors"].as_bool().unwrap_or(false),
-                (Timestamp::from(payload.svt), Timestamp::from(payload.stt)),
-                payload.cid,
-            )
-        } else {
-            // First page: parse the request and pin the snapshot. If no as_of
-            // was supplied, anchor at "now" so the whole scan is consistent.
-            let start_id = match NodeId::new(
-                args.get("start_node_id")
-                    .and_then(|v| v.as_u64())
-                    .unwrap_or(0),
-            ) {
-                Ok(id) => id,
-                Err(e) => return self.invalid_argument(&e.to_string()),
-            };
-            let temporal = match self.resolve_bitemporal_as_of(
-                &Self::arg_str(args, "as_of_valid_time"),
-                &Self::arg_str(args, "as_of_transaction_time"),
-            ) {
-                Ok(t) => t,
-                Err(result) => return result,
-            };
-            let snapshot = temporal.unwrap_or((now, now));
-            (
-                start_id,
-                Self::arg_str(args, "edge_label").unwrap_or_default(),
-                Self::arg_str(args, "direction").unwrap_or_else(|| "outgoing".into()),
-                args.get("depth")
-                    .and_then(|v| v.as_u64())
-                    .map(|d| d as usize)
-                    .unwrap_or(1)
-                    .min(MAX_TRAVERSAL_DEPTH),
-                Self::arg_limit(args),
-                0usize,
-                args.get("include_vectors")
-                    .and_then(|v| v.as_bool())
-                    .unwrap_or(false),
-                snapshot,
-                String::new(),
-            )
-        };
-
-        let (results, has_more) = self.run_traversal(
-            start_id,
-            &edge_label,
-            &direction,
-            depth,
-            limit,
-            offset,
-            Some(snapshot),
-            include_vectors,
-            now,
-        );
-        let count = results.len();
-
-        let mut obj = serde_json::Map::new();
-        obj.insert(
-            "results".to_string(),
-            serde_json::to_value(&results).unwrap_or_else(|_| json!([])),
-        );
-        obj.insert("count".to_string(), json!(count));
-        obj.insert(
-            "snapshot_valid_time".to_string(),
-            json!(Self::format_timestamp_rfc3339(snapshot.0)),
-        );
-        obj.insert(
-            "snapshot_transaction_time".to_string(),
-            json!(Self::format_timestamp_rfc3339(snapshot.1)),
-        );
-        obj.insert("has_more".to_string(), json!(has_more));
-        obj.insert("paging".to_string(), json!("cursor"));
-
-        if has_more {
-            let filters = json!({
-                "start_node_id": start_id.as_u64(),
-                "edge_label": edge_label,
-                "direction": direction,
-                "depth": depth,
-                "include_vectors": include_vectors,
-            });
-            let mut payload = CursorPayload::seed(
-                "traverse",
-                (snapshot.0.wallclock(), snapshot.1.wallclock()),
-                limit,
-                filters,
-            );
-            // Continuation offset advances by the number of rows ACTUALLY
-            // EMITTED (`count`). When #3353 token budgets land and can trim a
-            // page short of `limit`, `off` MUST advance by the post-trim row
-            // count (not `limit`), or the next page would skip trimmed rows.
-            payload.off = (offset + count) as u64;
-            payload.cid = parent_cid;
-            match self.cursors.issue(payload) {
-                Ok(token) => {
-                    obj.insert("cursor".to_string(), json!(token));
-                    obj.insert(
-                        "cursor_ttl_seconds".to_string(),
-                        json!(self.cursors.ttl().as_secs()),
-                    );
-                }
-                Err(e) => return self.error_result(e),
-            }
-        }
-
-        self.success_json(serde_json::Value::Object(obj))
-    }
-
-    fn handle_find_similar(&self, args: serde_json::Value) -> CallToolResult {
-        let req: FindSimilarRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        // Apply resource limits. A page must be able to carry a continuation
-        // cursor, so k is at least 1 (k:0 would otherwise report
-        // has_more:true with next_offset==offset, a non-progressing page).
-        let k = req.k.unwrap_or(DEFAULT_VECTOR_K).clamp(1, MAX_VECTOR_K);
-        // Bound the total pagination window (offset+k) by MAX_VECTOR_K so the
-        // over-fetch below never asks the vector index for more than
-        // MAX_VECTOR_K+1 candidates regardless of offset -- otherwise a large
-        // offset could force a search far past the MAX_VECTOR_K resource
-        // budget the cap is meant to enforce. Vector-similarity pagination is
-        // therefore bounded to the top MAX_VECTOR_K matches; an offset beyond
-        // that horizon returns an empty, complete (`has_more: false`) page.
-        let offset = req
-            .offset
-            .unwrap_or(0)
-            .min(MAX_PAGINATION_OFFSET)
-            .min(MAX_VECTOR_K.saturating_sub(k));
-
-        if !self.db.is_vector_index_enabled_for(&req.property_name) {
-            return self.failed_precondition(&format!(
-                "Vector index not enabled for property '{}'. Use enable_vector_index first.",
-                req.property_name
-            ));
-        }
-
-        // Validate embedding dimensions
-        if let Err(e) = self.validate_embedding_dimensions(&req.embedding, &req.property_name) {
-            return self.invalid_argument(&e);
-        }
-
-        // Over-fetch one past the requested page (offset + k + 1, capped at
-        // MAX_VECTOR_K + 1 by the offset bound above) so we can tell whether
-        // more similar nodes exist beyond this page (`has_more`) without a
-        // second query. The matching total would need a full index scan, so
-        // `total_matching` is omitted.
-        let fetch_k = k.saturating_add(offset).saturating_add(1);
-        match self
-            .db
-            .similarity_search(crate::SimilarityQuery::from_embedding(req.embedding).k(fetch_k))
-        {
-            Ok(results) => {
-                let has_more = results.len() > offset.saturating_add(k);
-                let include_vectors = req.include_vectors.unwrap_or(false);
-                // One request-scoped wallclock for every entity in the
-                // response (Issue #3391).
-                let now = time::now();
-                let similarity_results: Vec<SimilarityResult> = results
-                    .into_iter()
-                    .skip(offset)
-                    .take(k)
-                    .filter_map(|(node_id, score)| {
-                        self.db.get_node(node_id).ok().map(|node| SimilarityResult {
-                            node: self.node_to_response(&node, include_vectors, now),
-                            score,
-                        })
-                    })
-                    .collect();
-
-                let count = similarity_results.len();
-                let mut response = json!({
-                    "results": similarity_results,
-                    "count": count
-                });
-                // `next_offset` advances by the requested window `k`, not the
-                // (possibly smaller) resolved `count`: a since-deleted node
-                // behind a stale vector-index entry is still one of the `k`
-                // candidates this page consumed, so basing next_offset on
-                // `count` would re-skip into already-consumed candidates and
-                // duplicate a row on the next page.
-                Self::attach_completeness(&mut response, offset, k, has_more, None);
-                self.success_json(response)
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_enable_vector_index(&self, args: serde_json::Value) -> CallToolResult {
-        let req: EnableVectorIndexRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let distance_metric = match req.distance_metric.as_deref().unwrap_or("cosine") {
-            "euclidean" => DistanceMetric::Euclidean,
-            "dot" | "dot_product" => DistanceMetric::DotProduct,
-            _ => DistanceMetric::Cosine,
-        };
-
-        let config = HnswConfig::new(req.dimensions, distance_metric);
-
-        match self.db.enable_vector_index(&req.property_name, config) {
-            Ok(()) => self.success_json(json!({
-                "success": true,
-                "property_name": req.property_name,
-                "dimensions": req.dimensions,
-                "distance_metric": req.distance_metric.unwrap_or_else(|| "cosine".to_string())
-            })),
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_list_vector_indexes(&self, _args: serde_json::Value) -> CallToolResult {
-        let indexes = self.db.list_vector_indexes();
-        let index_list: Vec<serde_json::Value> = indexes
-            .into_iter()
-            .map(|info| {
-                json!({
-                    "property_name": info.property_name,
-                    "dimensions": info.dimensions,
-                    "distance_metric": format!("{:?}", info.distance_metric)
-                })
-            })
-            .collect();
-        self.success_json(json!({
-            "indexes": index_list,
-            "count": index_list.len()
-        }))
-    }
-
-    fn handle_enable_unique_constraint(&self, args: serde_json::Value) -> CallToolResult {
-        let req: EnableUniqueConstraintRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        match self
-            .db
-            .unique_constraint(&req.label, &req.property)
-            .enable()
-        {
-            Ok(()) => self.success_json(json!({
-                "success": true,
-                "label": req.label,
-                "property": req.property
-            })),
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_list_unique_constraints(&self, _args: serde_json::Value) -> CallToolResult {
-        let constraints = self.db.list_unique_constraints();
-        let list: Vec<serde_json::Value> = constraints
-            .into_iter()
-            .map(|(label, property)| json!({ "label": label, "property": property }))
-            .collect();
-        self.success_json(json!({
-            "constraints": list,
-            "count": list.len()
-        }))
-    }
-
-    fn handle_get_node_at_time(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetNodeAtTimeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let valid_time = match self.parse_timestamp(&req.valid_time) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&e),
-        };
-
-        let tx_time = match self.parse_optional_tx_time(req.transaction_time.as_deref()) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&e),
-        };
-
-        match self.db.get_node_at_time(node_id, valid_time, tx_time) {
-            Ok(node) => {
-                let now = time::now();
-                let response = self.node_to_response(&node, true, now);
-                self.success_json(json!({
-                    "node": response,
-                    "valid_time": req.valid_time,
-                    "transaction_time": Self::format_tx_time_response(req.transaction_time)
-                }))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_get_edge_at_time(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetEdgeAtTimeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let edge_id = match EdgeId::new(req.edge_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let valid_time = match self.parse_timestamp(&req.valid_time) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&e),
-        };
-
-        let tx_time = match self.parse_optional_tx_time(req.transaction_time.as_deref()) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&e),
-        };
-
-        match self.db.get_edge_at_time(edge_id, valid_time, tx_time) {
-            Ok(edge) => {
-                let now = time::now();
-                let response = self.edge_to_response(&edge, true, now);
-                self.success_json(json!({
-                    "edge": response,
-                    "valid_time": req.valid_time,
-                    "transaction_time": Self::format_tx_time_response(req.transaction_time)
-                }))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    /// Cursor-mode `find_nodes_at_time` (Issue #3360): snapshot-anchored keyset
-    /// paging. The snapshot is the caller's requested `(valid_time,
-    /// transaction_time)` -- already a point-in-time read, so consistency is
-    /// native; continuation just seeks by node id past the last returned.
-    fn handle_find_nodes_at_time_cursor(&self, args: &serde_json::Value) -> CallToolResult {
-        let now = time::now();
-
-        if let Some(token) = args.get("cursor").and_then(|v| v.as_str()) {
-            let payload = match self.cursors.decode(token, "find_nodes_at_time") {
-                Ok(p) => p,
-                Err(e) => return self.error_result(e),
-            };
-            let label = payload.filters["label"].as_str().unwrap_or("").to_string();
-            let property_key = payload.filters["property_key"].as_str().map(str::to_string);
-            let property_value = match &payload.filters["property_value"] {
-                serde_json::Value::Null => None,
-                v => Some(v.clone()),
-            };
-            let include_vectors = payload.filters["include_vectors"]
-                .as_bool()
-                .unwrap_or(false);
-            let snapshot = (Timestamp::from(payload.svt), Timestamp::from(payload.stt));
-            let candidates = match self.fetch_node_candidates(
-                &label,
-                &property_key,
-                &property_value,
-                snapshot,
-            ) {
-                Ok(c) => c,
-                Err(result) => return result,
-            };
-            return self.emit_node_cursor_page(
-                "find_nodes_at_time",
-                snapshot,
-                payload.after,
-                payload.limit,
-                include_vectors,
-                payload.filters.clone(),
-                candidates,
-                payload.cid,
-                now,
-            );
-        }
-
-        // First page: validate filter combo and pin the requested coordinate.
-        let label = Self::arg_str(args, "label").unwrap_or_default();
-        let property_key = Self::arg_str(args, "property_key");
-        let property_value = Self::arg_value(args, "property_value");
-        if property_key.is_some() != property_value.is_some() {
-            return self.invalid_argument(
-                "Both 'property_key' and 'property_value' are required together",
-            );
-        }
-        let valid_time_str = Self::arg_str(args, "valid_time").unwrap_or_default();
-        let valid_time = match self.parse_timestamp(&valid_time_str) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&format!("Invalid valid_time: {}", e)),
-        };
-        let tx_time = match self
-            .parse_optional_tx_time(Self::arg_str(args, "transaction_time").as_deref())
-        {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&format!("Invalid transaction_time: {}", e)),
-        };
-        let limit = Self::arg_limit(args);
-        let include_vectors = args
-            .get("include_vectors")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let snapshot = (valid_time, tx_time);
-        let candidates =
-            match self.fetch_node_candidates(&label, &property_key, &property_value, snapshot) {
-                Ok(c) => c,
-                Err(result) => return result,
-            };
-        let filters =
-            Self::node_scan_filters(&label, &property_key, &property_value, include_vectors);
-        self.emit_node_cursor_page(
-            "find_nodes_at_time",
-            snapshot,
-            None,
-            limit,
-            include_vectors,
-            filters,
-            candidates,
-            String::new(),
-            now,
-        )
-    }
-
-    /// Find nodes by label (and optional exact property match) as of a
-    /// bi-temporal point (Issue #3236).
-    ///
-    /// Validation mirrors `handle_list_nodes` (both-or-neither property
-    /// filter, same limit/offset clamps); the temporal reconstruction is
-    /// delegated to `AletheiaDB::find_nodes_at_time` /
-    /// `find_nodes_by_property_at`, which reconstruct each candidate from
-    /// the historical version visible at the queried coordinate -- so nodes
-    /// deleted from current state are still found when both dimensions
-    /// anchor before the deletion. The candidate set is capped at the same
-    /// `max_schema_as_of_entities` limit bi-temporal `get_schema` uses; when
-    /// truncated, the response discloses it via `sampled: true` and
-    /// `total_matching`/`has_more` count matches within the sampled
-    /// candidate set only.
-    fn handle_find_nodes_at_time(&self, args: serde_json::Value) -> CallToolResult {
-        // Snapshot-anchored cursor paging (Issue #3360); offset paging below
-        // is unchanged for backward compatibility.
-        if Self::cursor_requested(&args) {
-            return self.handle_find_nodes_at_time_cursor(&args);
-        }
-
-        let req: FindNodesAtTimeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        // Validate property filter: both key and value are required together
-        // (mirroring list_nodes).
-        if req.property_key.is_some() != req.property_value.is_some() {
-            return self.invalid_argument(
-                "Both 'property_key' and 'property_value' are required together",
-            );
-        }
-
-        let valid_time = match self.parse_timestamp(&req.valid_time) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&format!("Invalid valid_time: {}", e)),
-        };
-        let tx_time = match self.parse_optional_tx_time(req.transaction_time.as_deref()) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&format!("Invalid transaction_time: {}", e)),
-        };
-
-        // Apply resource limits exactly like list_nodes: a page must be able
-        // to carry a continuation cursor, so the limit is at least 1.
-        let limit = req
-            .limit
-            .unwrap_or(DEFAULT_RESULT_LIMIT)
-            .clamp(1, MAX_RESULT_LIMIT);
-        let offset = req.offset.unwrap_or(0).min(MAX_PAGINATION_OFFSET);
-
-        let matches =
-            if let (Some(prop_key), Some(prop_val)) = (&req.property_key, &req.property_value) {
-                let property_value = match self.json_to_property_value(prop_val) {
-                    Some(v) => v,
-                    None => return self.invalid_argument(
-                        "Unsupported property_value type. Use strings, numbers, booleans, or null.",
-                    ),
-                };
-                self.db.find_nodes_by_property_at(
-                    &req.label,
-                    prop_key,
-                    &property_value,
-                    valid_time,
-                    tx_time,
-                )
-            } else {
-                self.db.find_nodes_at_time(&req.label, valid_time, tx_time)
-            };
-
-        match matches {
-            Ok(matches) => {
-                // The matching set is already materialized (sorted by node
-                // id for stable pagination), so the total is cheap to report
-                // and `has_more` is exact *within the candidate set*. When
-                // `sampled` is true the candidate enumeration was truncated
-                // at the configured cap, so `total_matching` is honest only
-                // about the sampled candidates -- the flag discloses that.
-                let sampled = matches.sampled;
-                let matches = matches.nodes;
-                let total_matching = matches.len();
-                let include_vectors = req.include_vectors.unwrap_or(false);
-                // One request-scoped wallclock for every entity in the
-                // response (Issue #3391).
-                let now = time::now();
-                let nodes: Vec<NodeResponse> = matches
-                    .iter()
-                    .skip(offset)
-                    .take(limit)
-                    .map(|node| self.node_to_response(node, include_vectors, now))
-                    .collect();
-
-                let has_more = offset.saturating_add(limit) < total_matching;
-                let mut response = json!({
-                    "nodes": nodes,
-                    "count": nodes.len(),
-                    "offset": offset,
-                    "limit": limit,
-                    // Candidate-set truncation disclosure, mirroring
-                    // get_schema's `sampled` (same underlying cap).
-                    "sampled": sampled,
-                    // The resolved coordinate this answer holds at -- the
-                    // omitted transaction_time resolves to a concrete "now".
-                    "valid_time": Self::format_timestamp_rfc3339(valid_time),
-                    "transaction_time": Self::format_timestamp_rfc3339(tx_time),
-                });
-                Self::attach_completeness(
-                    &mut response,
-                    offset,
-                    limit,
-                    has_more,
-                    Some(total_matching),
-                );
-                self.success_json(response)
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_list_changes(&self, args: serde_json::Value) -> CallToolResult {
-        let req: ListChangesRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let tx_from = match self.parse_timestamp(&req.tx_from) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&format!("Invalid tx_from: {}", e)),
-        };
-        let tx_to = match self.parse_timestamp(&req.tx_to) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&format!("Invalid tx_to: {}", e)),
-        };
-
-        let valid_from = match self.parse_opt_timestamp("valid_from", &req.valid_from) {
-            Ok(v) => v,
-            Err(resp) => return resp,
-        };
-        let valid_to = match self.parse_opt_timestamp("valid_to", &req.valid_to) {
-            Ok(v) => v,
-            Err(resp) => return resp,
-        };
-
-        // A page must be able to carry a continuation cursor, so the limit is at least 1.
-        let limit = req
-            .limit
-            .unwrap_or(DEFAULT_RESULT_LIMIT)
-            .clamp(1, MAX_RESULT_LIMIT);
-
-        let query = ChangeFeedQuery {
-            tx_from,
-            tx_to,
-            valid_from,
-            valid_to,
-            label: req.label.clone(),
-            limit,
-            cursor: req.cursor.clone(),
-        };
-
-        match self.db.list_changes(&query) {
-            Ok(page) => {
-                let changes: Vec<serde_json::Value> = page
-                    .changes
-                    .iter()
-                    .map(|record| {
-                        json!({
-                            "entity_id": record.entity_id,
-                            "version_id": record.version_id,
-                            "kind": record.kind.as_str(),
-                            "change_type": record.change_type.as_str(),
-                            "label": record.label,
-                            "transaction_time": time::to_iso8601(record.transaction_time()),
-                            "transaction_time_range": {
-                                "start": time::to_iso8601(record.transaction_time_range.start()),
-                                "end": time::to_iso8601(record.transaction_time_range.end()),
-                            },
-                            "valid_time_range": {
-                                "start": time::to_iso8601(record.valid_time_range.start()),
-                                "end": time::to_iso8601(record.valid_time_range.end()),
-                            },
-                        })
-                    })
-                    .collect();
-
-                self.success_json(json!({
-                    "changes": changes,
-                    "count": page.changes.len(),
-                    "next_cursor": page.next_cursor,
-                }))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
     // ============================================================================
     // Phase 11: Independent Dimension Temporal Queries & Version History
     // ============================================================================
-
-    fn handle_get_node_at_valid_time(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetNodeAtValidTimeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let valid_time = match self.parse_timestamp(&req.valid_time) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&e),
-        };
-
-        match self.db.get_node_at_valid_time(node_id, valid_time) {
-            Ok(node) => {
-                let now = time::now();
-                let response = self.node_to_response(&node, true, now);
-                self.success_json(json!({
-                    "node": response,
-                    "valid_time": req.valid_time
-                }))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_get_node_at_transaction_time(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetNodeAtTransactionTimeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let tx_time = match self.parse_timestamp(&req.transaction_time) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&e),
-        };
-
-        match self.db.get_node_at_transaction_time(node_id, tx_time) {
-            Ok(node) => {
-                let now = time::now();
-                let response = self.node_to_response(&node, true, now);
-                self.success_json(json!({
-                    "node": response,
-                    "transaction_time": req.transaction_time
-                }))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_get_node_history(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetNodeHistoryRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        match self.db.get_node_history(node_id) {
-            Ok(history) => {
-                let versions: Vec<_> = history
-                    .versions
-                    .iter()
-                    .map(|v| self.version_info_to_response(v))
-                    .collect();
-
-                self.success_json(json!({
-                    "node_id": req.node_id,
-                    "versions": versions,
-                    "version_count": versions.len()
-                }))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_diff_node_versions(&self, args: serde_json::Value) -> CallToolResult {
-        let req: DiffNodeVersionsRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let node_id = match NodeId::new(req.node_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let from_version = match crate::core::id::VersionId::new(req.from_version) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let to_version = match crate::core::id::VersionId::new(req.to_version) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        match self
-            .db
-            .diff_node_versions(node_id, from_version, to_version)
-        {
-            Ok(diff) => {
-                let response = self.version_diff_to_response(&diff);
-                self.success_json(json!(response))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_get_edge_at_valid_time(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetEdgeAtValidTimeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let edge_id = match EdgeId::new(req.edge_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let valid_time = match self.parse_timestamp(&req.valid_time) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&e),
-        };
-
-        match self.db.get_edge_at_valid_time(edge_id, valid_time) {
-            Ok(edge) => {
-                let now = time::now();
-                let response = self.edge_to_response(&edge, true, now);
-                self.success_json(json!({
-                    "edge": response,
-                    "valid_time": req.valid_time
-                }))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_get_edge_at_transaction_time(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetEdgeAtTransactionTimeRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let edge_id = match EdgeId::new(req.edge_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let tx_time = match self.parse_timestamp(&req.transaction_time) {
-            Ok(t) => t,
-            Err(e) => return self.invalid_argument(&e),
-        };
-
-        match self.db.get_edge_at_transaction_time(edge_id, tx_time) {
-            Ok(edge) => {
-                let now = time::now();
-                let response = self.edge_to_response(&edge, true, now);
-                self.success_json(json!({
-                    "edge": response,
-                    "transaction_time": req.transaction_time
-                }))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_get_edge_history(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetEdgeHistoryRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let edge_id = match EdgeId::new(req.edge_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        match self.db.get_edge_history(edge_id) {
-            Ok(history) => {
-                let versions: Vec<_> = history
-                    .versions
-                    .iter()
-                    .map(|v| self.version_info_to_response(v))
-                    .collect();
-
-                self.success_json(json!({
-                    "edge_id": req.edge_id,
-                    "versions": versions,
-                    "version_count": versions.len()
-                }))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    fn handle_diff_edge_versions(&self, args: serde_json::Value) -> CallToolResult {
-        let req: DiffEdgeVersionsRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let edge_id = match EdgeId::new(req.edge_id) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let from_version = match crate::core::id::VersionId::new(req.from_version) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        let to_version = match crate::core::id::VersionId::new(req.to_version) {
-            Ok(id) => id,
-            // An out-of-range ID is a caller fault; emit the bare
-            // `StorageError` text verbatim (`db_error` would wrap it in
-            // `Error::Storage`, prefixing "Storage error: " — a message
-            // regression vs pre-#3234 responses).
-            Err(e) => return self.invalid_argument(&e.to_string()),
-        };
-
-        match self
-            .db
-            .diff_edge_versions(edge_id, from_version, to_version)
-        {
-            Ok(diff) => {
-                let response = self.version_diff_to_response(&diff);
-                self.success_json(json!(response))
-            }
-            Err(e) => self.db_error(e),
-        }
-    }
 
     // Helper methods for converting internal types to response types
 
@@ -4190,33 +2066,6 @@ impl AletheiaMcpServer {
         })
     }
 
-    /// Discover the graph's schema (labels, edge types, property keys),
-    /// optionally as of a bi-temporal instant. Never errors on an empty
-    /// database — returns a well-formed, empty summary instead.
-    fn handle_get_schema(&self, args: serde_json::Value) -> CallToolResult {
-        let req: GetSchemaRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let temporal = match self
-            .resolve_bitemporal_as_of(&req.as_of_valid_time, &req.as_of_transaction_time)
-        {
-            Ok(t) => t,
-            Err(result) => return result,
-        };
-
-        let result = match temporal {
-            None => self.db.schema(),
-            Some((vt, tt)) => self.db.schema_as_of(vt, tt),
-        };
-
-        match result {
-            Ok(schema) => self.success_json(self.schema_to_json(&schema)),
-            Err(e) => self.db_error(e),
-        }
-    }
-
     /// Render a finite timestamp as an RFC3339 string (UTC, microsecond
     /// precision). Falls back to a raw microsecond form for coordinates
     /// outside chrono's representable range rather than panicking.
@@ -4274,340 +2123,6 @@ impl AletheiaMcpServer {
         }
 
         value
-    }
-
-    /// Report the dataset's queryable bi-temporal extent (Issue #3238).
-    ///
-    /// The handler only serializes: bounds come from the public
-    /// `AletheiaDB::temporal_extent` / `temporal_extent_by_label` API.
-    fn handle_temporal_extent(&self, args: serde_json::Value) -> CallToolResult {
-        // The tool has no required arguments: a call with no `arguments`
-        // object at all must behave like `{}`.
-        let args = if args.is_null() {
-            serde_json::Value::Object(serde_json::Map::new())
-        } else {
-            args
-        };
-
-        let req: TemporalExtentRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let result = if req.by_label.unwrap_or(false) {
-            self.db.temporal_extent_by_label()
-        } else {
-            self.db.temporal_extent()
-        };
-
-        match result {
-            Ok(extent) => self.success_json(Self::temporal_extent_to_json(&extent)),
-            Err(e) => self.db_error(e),
-        }
-    }
-
-    /// Handle the `audit_export` tool (Issue #3358).
-    ///
-    /// Produces a signed, offline-verifiable evidence artifact of an entity's
-    /// complete bi-temporal history. The Ed25519 signing key is operator-
-    /// provided out of band via the `ALETHEIADB_AUDIT_SIGNING_KEY` environment
-    /// variable (a 32-byte hex seed); the secret is never returned or logged —
-    /// only the public key travels in the artifact.
-    fn handle_audit_export(&self, args: serde_json::Value) -> CallToolResult {
-        use crate::audit::{AuditScope, AuditSigningKey, ExportOptions, SIGNING_KEY_ENV};
-
-        let req: AuditExportRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        let scope = match req.entity_type.as_str() {
-            "node" => match NodeId::new(req.entity_id) {
-                Ok(id) => AuditScope::node(id),
-                Err(e) => return self.invalid_argument(&e.to_string()),
-            },
-            "edge" => match crate::core::id::EdgeId::new(req.entity_id) {
-                Ok(id) => AuditScope::edge(id),
-                Err(e) => return self.invalid_argument(&e.to_string()),
-            },
-            other => {
-                return self.invalid_argument(&format!(
-                    "entity_type must be 'node' or 'edge', got '{other}'"
-                ));
-            }
-        };
-
-        // The signing key is a precondition supplied by the operator, not a
-        // caller argument — a missing key is a FAILED_PRECONDITION, never a
-        // silent unsigned export.
-        let signing_key = match AuditSigningKey::from_env(SIGNING_KEY_ENV) {
-            Ok(k) => k,
-            Err(_) => {
-                return self.failed_precondition(&format!(
-                    "audit export requires an operator-provided Ed25519 signing key in the \
-                     {SIGNING_KEY_ENV} environment variable (32-byte hex seed)"
-                ));
-            }
-        };
-
-        let mut options =
-            ExportOptions::new(req.database_id.unwrap_or_else(|| "aletheiadb".to_string()));
-        if !req.redact_keys.is_empty() {
-            options = options.redact(req.redact_keys);
-        }
-
-        match self.db.audit_export(scope, &signing_key, &options) {
-            Ok(export) => match serde_json::to_value(&export) {
-                Ok(artifact) => self.success_json(json!({
-                    "artifact": artifact,
-                    "public_key": signing_key.public_key().to_hex(),
-                    "entity_count": export.entity_count(),
-                    "version_count": export.version_count(),
-                    "chain_root": export.chain.root,
-                })),
-                Err(e) => self.error_result(McpError::new(
-                    McpErrorCode::Internal,
-                    format!("failed to serialize artifact: {e}"),
-                )),
-            },
-            Err(crate::audit::AuditError::NoHistory(msg)) => self.error_result(McpError::new(
-                McpErrorCode::NotFound,
-                format!("no exportable history: {msg}"),
-            )),
-            Err(e) => self.error_result(McpError::new(
-                McpErrorCode::Internal,
-                format!("audit export failed: {e}"),
-            )),
-        }
-    }
-
-    /// Handle the `database_stats` tool (Issue #3222).
-    ///
-    /// Thin aggregator: delegates entirely to the public
-    /// [`AletheiaDB::stats`] snapshot and serializes it — no storage logic
-    /// lives here. The underlying getters are all O(1)/cached (see
-    /// `src/db/stats.rs`), so this never triggers a version scan.
-    fn handle_database_stats(&self, args: serde_json::Value) -> CallToolResult {
-        // The tool takes no required arguments; clients may send no
-        // `arguments` at all (surfaced here as JSON null) or an empty
-        // object. Normalize null so both forms are accepted.
-        let args = if args.is_null() {
-            serde_json::Value::Object(serde_json::Map::new())
-        } else {
-            args
-        };
-        let _req: DatabaseStatsRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        match serde_json::to_value(self.db.stats()) {
-            Ok(value) => self.success_json(value),
-            Err(e) => self.error_result(McpError::new(
-                McpErrorCode::Internal,
-                format!("Failed to serialize database stats: {}", e),
-            )),
-        }
-    }
-
-    fn handle_hybrid_query(&self, args: serde_json::Value) -> CallToolResult {
-        let req: HybridQueryRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => return self.invalid_argument(&format!("Invalid arguments: {}", e)),
-        };
-
-        // Apply resource limits
-        let limit = req
-            .limit
-            .unwrap_or(DEFAULT_RESULT_LIMIT)
-            .min(MAX_RESULT_LIMIT);
-        let depth = req.traverse_depth.unwrap_or(1).min(MAX_TRAVERSAL_DEPTH);
-        let k = req.top_k.unwrap_or(DEFAULT_VECTOR_K).min(MAX_VECTOR_K);
-
-        // Parse temporal parameters if provided
-        let valid_time = if let Some(ref vt) = req.valid_time {
-            match self.parse_timestamp(vt) {
-                Ok(t) => Some(t),
-                Err(e) => return self.invalid_argument(&format!("Invalid valid_time: {}", e)),
-            }
-        } else {
-            None
-        };
-
-        let tx_time = if let Some(ref tt) = req.transaction_time {
-            match self.parse_timestamp(tt) {
-                Ok(t) => Some(t),
-                Err(e) => {
-                    return self.invalid_argument(&format!("Invalid transaction_time: {}", e));
-                }
-            }
-        } else {
-            None
-        };
-
-        let include_vectors = req.include_vectors.unwrap_or(false);
-
-        // One request-scoped wallclock for every entity in the response
-        // (Issue #3391).
-        let now = time::now();
-
-        // Helper to convert rows to hybrid results with temporal info
-        let rows_to_results =
-            |rows: Vec<crate::query::executor::QueryRow>| -> Vec<HybridQueryResult> {
-                rows.into_iter()
-                    .filter_map(|row| {
-                        if let EntityResult::Node(node) = row.entity {
-                            Some(HybridQueryResult {
-                                node: self.node_to_response(&node, include_vectors, now),
-                                similarity_score: row.score,
-                                traversal_path: row.path.map(|p| {
-                                    p.iter()
-                                        .map(|e| match e {
-                                            ResultEntityId::Node(id) => id.as_u64(),
-                                            ResultEntityId::Edge(id) => id.as_u64(),
-                                        })
-                                        .collect()
-                                }),
-                                timestamp: row.timestamp.map(|t| t.wallclock().to_string()),
-                            })
-                        } else {
-                            None
-                        }
-                    })
-                    .collect()
-            };
-
-        // Use QueryBuilder for hybrid queries
-        if let Some(start_id) = req.start_node_id {
-            let node_id = match NodeId::new(start_id) {
-                Ok(id) => id,
-                // Bare StorageError text verbatim — see the note on the other
-                // ID-validation sites.
-                Err(e) => return self.invalid_argument(&e.to_string()),
-            };
-
-            // If temporal filtering requested, use temporal query
-            if let (Some(vt), Some(tt)) = (valid_time, tx_time) {
-                // Temporal query for a single node
-                return match self.db.get_node_at_time(node_id, vt, tt) {
-                    Ok(node) => {
-                        let response = self.node_to_response(&node, include_vectors, now);
-                        self.success_json(json!({
-                            "results": [HybridQueryResult {
-                                node: response,
-                                similarity_score: None,
-                                traversal_path: Some(vec![node_id.as_u64()]),
-                                timestamp: Some(vt.wallclock().to_string()),
-                            }],
-                            "count": 1,
-                            "temporal_query": {
-                                "valid_time": req.valid_time,
-                                "transaction_time": req.transaction_time
-                            }
-                        }))
-                    }
-                    Err(e) => self.db_error(e),
-                };
-            }
-
-            // Graph-first query with optional vector ranking
-            let builder = crate::query::QueryBuilder::new().start(node_id);
-
-            let builder = if let Some(ref edge_label) = req.traverse_edge {
-                if depth > 1 {
-                    builder.traverse_n(edge_label, depth)
-                } else {
-                    builder.traverse(edge_label)
-                }
-            } else {
-                // Just return the start node
-                return match self.db.get_node(node_id) {
-                    Ok(node) => {
-                        let response = self.node_to_response(&node, include_vectors, now);
-                        self.success_json(json!({
-                            "results": [HybridQueryResult {
-                                node: response,
-                                similarity_score: None,
-                                traversal_path: Some(vec![node_id.as_u64()]),
-                                timestamp: None,
-                            }],
-                            "count": 1
-                        }))
-                    }
-                    Err(e) => self.db_error(e),
-                };
-            };
-
-            // Execute and collect results
-            match builder.limit(limit).execute(&self.db) {
-                Ok(results) => match results.collect_all() {
-                    Ok(rows) => {
-                        let hybrid_results = rows_to_results(rows);
-                        self.success_json(json!({
-                            "results": hybrid_results,
-                            "count": hybrid_results.len()
-                        }))
-                    }
-                    Err(e) => self.db_error(e),
-                },
-                Err(e) => self.db_error(e),
-            }
-        } else if let Some(ref embedding) = req.query_embedding {
-            // Vector-first query
-            // Use vector_property if specified
-            let property_name = req.vector_property.as_deref().unwrap_or("embedding");
-
-            // Check if vector index is enabled for the property
-            if !self.db.is_vector_index_enabled_for(property_name) {
-                return self.failed_precondition(&format!(
-                    "Vector index not enabled for property '{}'. Use enable_vector_index first.",
-                    property_name
-                ));
-            }
-
-            // Validate embedding dimensions
-            if let Err(e) = self.validate_embedding_dimensions(embedding, property_name) {
-                return self.invalid_argument(&e);
-            }
-
-            let builder = crate::query::QueryBuilder::new().find_similar(embedding, k);
-
-            match builder.limit(limit).execute(&self.db) {
-                Ok(results) => match results.collect_all() {
-                    Ok(rows) => {
-                        let hybrid_results = rows_to_results(rows);
-                        self.success_json(json!({
-                            "results": hybrid_results,
-                            "count": hybrid_results.len(),
-                            "vector_property": property_name
-                        }))
-                    }
-                    Err(e) => self.db_error(e),
-                },
-                Err(e) => self.db_error(e),
-            }
-        } else if let Some(ref label) = req.filter_label {
-            // Label scan query
-            let builder = crate::query::QueryBuilder::new().scan_label(label);
-
-            match builder.limit(limit).execute(&self.db) {
-                Ok(results) => match results.collect_all() {
-                    Ok(rows) => {
-                        let hybrid_results = rows_to_results(rows);
-                        self.success_json(json!({
-                            "results": hybrid_results,
-                            "count": hybrid_results.len()
-                        }))
-                    }
-                    Err(e) => self.db_error(e),
-                },
-                Err(e) => self.db_error(e),
-            }
-        } else {
-            self.invalid_argument(
-                "Must specify either start_node_id, query_embedding, or filter_label",
-            )
-        }
     }
 
     // ========================================================================
@@ -4812,152 +2327,6 @@ impl AletheiaMcpServer {
             out.insert(key.clone(), pv);
         }
         Ok(out)
-    }
-
-    fn handle_query(&self, args: serde_json::Value) -> CallToolResult {
-        // Extract language early so it can appear in error payloads even when
-        // full deserialization fails (language is not yet known at that point).
-        let raw_language = args
-            .get("language")
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_ascii_lowercase());
-
-        // Cursor paging is not supported for the declarative query tool in v1
-        // (Issue #3360); captured before `args` is consumed by deserialization.
-        let cursor_requested = Self::cursor_requested(&args);
-
-        let req: QueryRequest = match serde_json::from_value(args) {
-            Ok(r) => r,
-            Err(e) => {
-                return self.query_error(
-                    "invalid_request",
-                    &format!("Invalid arguments: {e}"),
-                    None,
-                    raw_language.as_deref(),
-                );
-            }
-        };
-
-        let language = req.language.to_ascii_lowercase();
-        if language != "cypher" && language != "aql" {
-            return self.query_error(
-                "invalid_request",
-                &format!(
-                    "Unsupported query language '{}'. Use \"cypher\" or \"aql\".",
-                    req.language
-                ),
-                None,
-                Some(&language),
-            );
-        }
-
-        // Cursor paging (Issue #3360) is not supported for the declarative
-        // query tool in v1: arbitrary result shapes (projections, aggregates,
-        // ordering) have no snapshot-anchored keyset to page over. Return a
-        // structured `unsupported_construct` error rather than silently
-        // serving a single truncated page (AC7: no silent fallback). Callers
-        // needing consistent, resumable scans use `list_nodes` /
-        // `find_nodes_at_time`, which are cursor-paged.
-        if cursor_requested {
-            return self.query_error(
-                "unsupported_construct",
-                "Cursor paging is not supported for the `query` tool in v1. Use `list_nodes` or \
-                 `find_nodes_at_time` for snapshot-anchored, resumable cursor scans; the `query` \
-                 tool returns a single (optionally `limit`-bounded) result set.",
-                None,
-                Some(&language),
-            );
-        }
-
-        // Read-only guard: reject mutating statements BEFORE any execution.
-        // Runs for every language so the tool can never write, even if the
-        // grammars later gain write support.
-        if let Some(clause) = detect_mutating_clause(&req.query) {
-            return self.query_error(
-                "read_only_violation",
-                &format!(
-                    "The `query` tool is read-only; the `{clause}` clause would mutate state and \
-                     is rejected before execution."
-                ),
-                Some(clause),
-                Some(&language),
-            );
-        }
-
-        let limit = req
-            .limit
-            .unwrap_or(DEFAULT_RESULT_LIMIT)
-            .min(MAX_RESULT_LIMIT);
-        let has_params = req.params.as_ref().is_some_and(|p| !p.is_empty());
-
-        let execution = match language.as_str() {
-            "aql" => {
-                if has_params {
-                    return self.query_error(
-                        "invalid_request",
-                        "AQL does not support parameter bindings; inline literal values or use \
-                         language \"cypher\" with $params.",
-                        None,
-                        Some("aql"),
-                    );
-                }
-                self.db.execute_aql(&req.query)
-            }
-            "cypher" => {
-                #[cfg(feature = "cypher")]
-                {
-                    match self.json_to_cypher_params(req.params.as_ref()) {
-                        Ok(params) if params.is_empty() => self.db.execute_cypher(&req.query),
-                        Ok(params) => self.db.execute_cypher_with_params(&req.query, params),
-                        Err((parameter, reason)) => {
-                            return self.query_error(
-                                "invalid_params",
-                                &format!("Invalid parameter '{parameter}': {reason}"),
-                                None,
-                                Some("cypher"),
-                            );
-                        }
-                    }
-                }
-                #[cfg(not(feature = "cypher"))]
-                {
-                    return self.query_error(
-                        "language_unavailable",
-                        "Cypher support is not compiled in (enable the `cypher` feature). Use \
-                         language \"aql\" instead.",
-                        None,
-                        Some("cypher"),
-                    );
-                }
-            }
-            _ => unreachable!("language already validated above"),
-        };
-
-        let results = match execution {
-            Ok(results) => results,
-            Err(e) => return self.map_query_error(e, &language),
-        };
-
-        // Collect one extra row to detect (and report) truncation at the cap.
-        let collected = match results.take_n(limit.saturating_add(1)) {
-            Ok(rows) => rows,
-            Err(e) => return self.map_query_error(e, &language),
-        };
-        let truncated = collected.len() > limit;
-        let rows: Vec<serde_json::Value> = collected
-            .into_iter()
-            .take(limit)
-            .map(|row| self.query_row_to_json(row))
-            .collect();
-        let row_count = rows.len();
-
-        self.success_json(json!({
-            "language": language,
-            "columns": query_columns(),
-            "rows": rows,
-            "row_count": row_count,
-            "truncated": truncated,
-        }))
     }
 
     /// Test-only accessor returning the names of all advertised tools.
@@ -5252,6 +2621,17 @@ fn query_columns() -> serde_json::Value {
         })
         .clone()
 }
+
+mod audit;
+mod edges;
+mod hybrid;
+mod nodes;
+mod schema;
+mod temporal;
+#[cfg(test)]
+mod tests;
+mod traversal;
+mod vector;
 
 /// Build the list of tool definitions advertised by this MCP server.
 ///
@@ -5800,197 +3180,5 @@ impl ServerHandler for AletheiaMcpServer {
             .unwrap_or(serde_json::Value::Null);
 
         Ok(self.dispatch_tool(request.name.as_ref(), args))
-    }
-}
-
-#[cfg(test)]
-mod server_unit_tests {
-    use std::sync::Arc;
-
-    use super::AletheiaMcpServer;
-    use crate::core::error::{Error, QueryError};
-    use crate::core::id::{EdgeId, NodeId};
-    use crate::db::AletheiaDB;
-    use crate::query::executor::{EntityResult, QueryRow};
-
-    fn make_server() -> AletheiaMcpServer {
-        AletheiaMcpServer::new(Arc::new(AletheiaDB::new().expect("db init")))
-    }
-
-    fn error_kind(server: &AletheiaMcpServer, err: Error) -> String {
-        let result = server.map_query_error(err, "aql");
-        let text = AletheiaMcpServer::extract_text(result);
-        let val: serde_json::Value = serde_json::from_str(&text).unwrap();
-        val["error"]["kind"].as_str().unwrap_or("").to_string()
-    }
-
-    /// Like [`error_kind`], but returning the whole serialized `error` object
-    /// so tests can assert `code`/`retriable` alongside `kind`.
-    fn error_payload(server: &AletheiaMcpServer, err: Error) -> serde_json::Value {
-        let result = server.map_query_error(err, "aql");
-        let text = AletheiaMcpServer::extract_text(result);
-        let val: serde_json::Value = serde_json::from_str(&text).unwrap();
-        val["error"].clone()
-    }
-
-    #[test]
-    fn map_query_error_unsupported_feature_yields_unsupported_construct() {
-        let server = make_server();
-        let err = Error::Query(QueryError::UnsupportedFeature {
-            feature: "DISTINCT".to_string(),
-        });
-        assert_eq!(error_kind(&server, err), "unsupported_construct");
-    }
-
-    #[test]
-    fn map_query_error_invalid_parameter_yields_invalid_params() {
-        let server = make_server();
-        let err = Error::Query(QueryError::InvalidParameter {
-            parameter: "p".to_string(),
-            reason: "out of range".to_string(),
-        });
-        assert_eq!(error_kind(&server, err), "invalid_params");
-    }
-
-    #[test]
-    fn map_query_error_execution_error_yields_runtime_error() {
-        let server = make_server();
-        let err = Error::Query(QueryError::ExecutionError {
-            message: "boom".to_string(),
-        });
-        assert_eq!(error_kind(&server, err), "runtime_error");
-    }
-
-    #[test]
-    fn map_query_error_other_variant_yields_runtime_error() {
-        let server = make_server();
-        // Error::Other is a variant not matched by any specific arm — falls through to `other`.
-        let err = Error::Other("unexpected situation".to_string());
-        assert_eq!(error_kind(&server, err), "runtime_error");
-    }
-
-    #[test]
-    fn map_query_error_timeout_yields_retriable_unavailable_runtime_error() {
-        // A timeout keeps the query tool's own `kind` contract
-        // ("runtime_error") but is classified UNAVAILABLE/retriable from the
-        // underlying engine error — and `retriable: true` must survive the
-        // query-tool serialization path, not just the in-memory struct.
-        let server = make_server();
-        let error = error_payload(
-            &server,
-            Error::Query(QueryError::Timeout { duration_ms: 5000 }),
-        );
-        assert_eq!(error["kind"], "runtime_error", "got: {error}");
-        assert_eq!(error["code"], "UNAVAILABLE", "got: {error}");
-        assert_eq!(error["retriable"], true, "got: {error}");
-    }
-
-    #[test]
-    fn query_row_to_json_node_id_variant() {
-        let server = make_server();
-        let row = QueryRow::from_entity(EntityResult::NodeId(NodeId::new(42).unwrap()));
-        let json = server.query_row_to_json(row);
-        assert_eq!(json["entity"]["type"].as_str(), Some("node"));
-        assert_eq!(json["entity"]["id"].as_u64(), Some(42));
-    }
-
-    #[test]
-    fn query_row_to_json_edge_id_variant() {
-        let server = make_server();
-        let row = QueryRow::from_entity(EntityResult::EdgeId(EdgeId::new(99).unwrap()));
-        let json = server.query_row_to_json(row);
-        assert_eq!(json["entity"]["type"].as_str(), Some("edge"));
-        assert_eq!(json["entity"]["id"].as_u64(), Some(99));
-    }
-
-    #[test]
-    fn handle_enable_unique_constraint_invalid_json_returns_error() {
-        // Covers the `Err(e) => return self.invalid_argument(...)` parse-error arm of
-        // handle_enable_unique_constraint (added for Issue #3218).  The public
-        // `enable_unique_constraint(req)` API always serialises a valid struct,
-        // so this arm is only reachable via the internal handle_ function.
-        let server = make_server();
-        let result = server.handle_enable_unique_constraint(serde_json::Value::Null);
-        // Must be an error CallToolResult (is_error = Some(true))
-        assert!(
-            result.is_error.unwrap_or(false),
-            "Null JSON input must produce an error result"
-        );
-    }
-
-    #[test]
-    fn timestamp_to_rfc3339_out_of_chrono_range_falls_back_to_micros() {
-        // Coordinates outside chrono's representable range must render as
-        // the raw-microseconds fallback instead of panicking. Timestamps up
-        // to MAX_VALID_TIMESTAMP (i64::MAX - 1000 µs) are storable but far
-        // beyond chrono's ~year-262143 ceiling.
-        let ts = crate::core::temporal::Timestamp::from(i64::MAX - 1000);
-        let rendered = AletheiaMcpServer::timestamp_to_rfc3339(ts);
-        assert_eq!(rendered, format!("{}us", i64::MAX - 1000));
-
-        // Sanity: an in-range coordinate renders as RFC3339, not the fallback.
-        let ts = crate::core::temporal::Timestamp::from(1_614_556_800_000_000); // 2021-03-01
-        let rendered = AletheiaMcpServer::timestamp_to_rfc3339(ts);
-        assert_eq!(rendered, "2021-03-01T00:00:00.000000Z");
-    }
-
-    #[test]
-    fn handle_temporal_extent_invalid_by_label_type_routes_through_invalid_argument() {
-        // A mistyped argument (string instead of bool) must produce the
-        // structured Issue #3234 error payload with code INVALID_ARGUMENT
-        // (Issue #3238).
-        let server = make_server();
-        let result = server.handle_temporal_extent(serde_json::json!({"by_label": "yes"}));
-        assert!(
-            result.is_error.unwrap_or(false),
-            "mistyped by_label must produce an error result"
-        );
-        let text = AletheiaMcpServer::extract_text(result);
-        let val: serde_json::Value = serde_json::from_str(&text).unwrap();
-        assert_eq!(
-            val["error"]["code"], "INVALID_ARGUMENT",
-            "mistyped by_label must classify as INVALID_ARGUMENT: {val}"
-        );
-        assert_eq!(
-            val["error"]["retriable"], false,
-            "INVALID_ARGUMENT must not be retriable: {val}"
-        );
-        assert!(
-            val["error"]["message"]
-                .as_str()
-                .is_some_and(|m| m.contains("Invalid arguments")),
-            "message must preserve the free-text detail: {val}"
-        );
-    }
-
-    #[test]
-    fn handle_temporal_extent_null_args_behaves_like_no_arguments() {
-        // The tool has no required arguments: an MCP call with the
-        // `arguments` object omitted entirely (routed here as Null) must
-        // succeed exactly like `{}`.
-        let server = make_server();
-        let result = server.handle_temporal_extent(serde_json::Value::Null);
-        assert!(
-            !result.is_error.unwrap_or(false),
-            "temporal_extent with no arguments must succeed"
-        );
-        let text = AletheiaMcpServer::extract_text(result);
-        let val: serde_json::Value = serde_json::from_str(&text).unwrap();
-        assert!(val["valid_time"]["earliest"].is_null());
-        assert!(val["transaction_time"]["latest"].is_null());
-    }
-
-    #[test]
-    fn query_row_to_json_null_binding_serializes_entity_as_json_null() {
-        // A null binding from an unmatched OPTIONAL MATCH pattern must
-        // surface as an explicit JSON null entity (row preserved).
-        let server = make_server();
-        let value = server.query_row_to_json(QueryRow::from_entity(EntityResult::Null));
-        assert!(
-            value["entity"].is_null(),
-            "null binding must serialize as JSON null: {value}"
-        );
-        assert!(value["score"].is_null());
-        assert!(value["path"].is_null());
     }
 }
